refactor: remove the `exit` module
